Extracting inner class to a new file
